Make signature cache optional
